+++ src/index.js
@@ -1,3 +1,5 @@
 export { default as Test } from './atoms/test';
 export { default as MyLife } from './atoms/MyLife';
 export { default as Transport } from './atoms/Transport';
+export { default as SocialIcon } from './atoms/SocialIcon';
+export { default as UserPhoto } from './atoms/UserPhoto';