+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v0.1.9
+
+- Adds User Photo Component
+- Adds Social Icons Component
+
 ## v0.1.8
 
 -adds transport icon